ESSIFI-109 Create a wrapper function for evaluate function
   - Moving the evaluateWrapper function to EvaluationClientWrapper class;
