+++ lib/evaluation/evaluationClient.ts
@@ -3,7 +3,6 @@ import { PresentationDefinition } from '@sphereon/pe-models';
 import { Checked, Status } from '../ConstraintUtils';
 
 import { EvaluationHandler } from './evaluationHandler';
-import { EvaluationResults } from './evaluationResults';
 import { HandlerCheckResult } from './handlerCheckResult';
 import { InputDescriptorFilterEvaluationHandler } from './inputDescriptorFilterEvaluationHandler';
 import { LimitDisclosureEvaluationHandler } from './limitDisclosureEvaluationHandler';
@@ -40,24 +39,6 @@ export class EvaluationClient {
       throw this.failed_catched;
     }
     return this._results;
-  }
-
-  public evaluateWrapper(pd: PresentationDefinition, vp: unknown): EvaluationResults {
-    this.evaluate(pd, vp);
-    const result: any = {};
-    result.warnings = this.results.filter((result) => result.status === Status.WARN).map((x) => JSON.stringify(x));
-    result.errors = this.results
-      .filter((result) => result.status === Status.ERROR)
-      .map((x) => {
-        return {
-          name: x.evaluator,
-          message: `${x.message}: ${x.input_descriptor_path}: ${x.verifiable_credential_path}`,
-        };
-      });
-    if (this._verifiablePresentation['presentationSubmission']['descriptor_map'].length) {
-      result.value = this._verifiablePresentation['presentationSubmission'];
-    }
-    return result;
   }
 
   public get results(): HandlerCheckResult[] {
