+++ lib/evaluation/warn.ts
@@ -1,0 +1,1 @@
+export class Warn extends Error {}
