+++ test/evaluation/evaluationClient.spec.ts
@@ -4,6 +4,7 @@ import { Optionality, PresentationDefinition } from '@sphereon/pe-models';
 
 import { Status } from '../../lib';
 import { EvaluationClient } from "../../lib/evaluation/evaluationClient";
+import { EvaluationClientWrapper } from '../../lib/evaluation/evaluationClientWrapper';
 
 function getFile(path: string) {
   return JSON.parse(fs.readFileSync(path, 'utf-8'));
@@ -106,7 +107,7 @@ describe('evaluate', () => {
     const vpSimple = getFile('./test/dif_pe_examples/vp/vp-simple-age-predicate.json');
     const evaluationClient: EvaluationClient = new EvaluationClient();
     pdSchema.input_descriptors[0].schema[0].uri = "https://www.w3.org/TR/vc-data-model/#types1";
-    const evaluationResults = evaluationClient.evaluateWrapper(pdSchema, vpSimple);
+    const evaluationResults = new EvaluationClientWrapper().evaluate(pdSchema, vpSimple, evaluationClient);
     expect(evaluationClient.results[0]).toEqual({
       "input_descriptor_path": "$.input_descriptors[0]",
       "verifiable_credential_path": "$.verifiableCredential[0]",
@@ -142,7 +143,7 @@ describe('evaluate', () => {
     const vpSimple = getFile('./test/dif_pe_examples/vp/vp-simple-age-predicate.json');
     const evaluationClient: EvaluationClient = new EvaluationClient();
     pdSchema.input_descriptors[0].schema.push({ uri: "https://www.w3.org/TR/vc-data-model/#types1" });
-    const evaluationResults = evaluationClient.evaluateWrapper(pdSchema, vpSimple);
+    const evaluationResults = new EvaluationClientWrapper().evaluate(pdSchema, vpSimple, evaluationClient);
     const errorResults = evaluationClient.results.filter(result => result.status === Status.ERROR);
     expect(errorResults.length).toEqual(0);
     expect(evaluationResults).toEqual(success);
@@ -153,7 +154,7 @@ describe('evaluate', () => {
     const vpSimple = getFile('./test/dif_pe_examples/vp/vp-simple-age-predicate.json');
     const evaluationClient: EvaluationClient = new EvaluationClient();
     vpSimple.verifiableCredential[0].credentialSchema[0].id = "https://www.w3.org/TR/vc-data-model/#types1";
-    const evaluationResults = evaluationClient.evaluateWrapper(pdSchema, vpSimple);
+    const evaluationResults = new EvaluationClientWrapper().evaluate(pdSchema, vpSimple, evaluationClient);
     expect(evaluationClient.results[0]).toEqual({
       "input_descriptor_path": "$.input_descriptors[0]",
       "verifiable_credential_path": "$.verifiableCredential[0]",
@@ -189,7 +190,7 @@ describe('evaluate', () => {
     const vpSimple = getFile('./test/dif_pe_examples/vp/vp-simple-age-predicate.json');
     const evaluationClient: EvaluationClient = new EvaluationClient();
     vpSimple.verifiableCredential[0].credentialSchema.push({ id: "https://www.w3.org/TR/vc-data-model/#types1" });
-    const evaluationResults = evaluationClient.evaluateWrapper(pdSchema, vpSimple);
+    const evaluationResults = new EvaluationClientWrapper().evaluate(pdSchema, vpSimple, evaluationClient);
     const errorResults = evaluationClient.results.filter(result => result.status === Status.ERROR);
     expect(errorResults.length).toEqual(2);
     expect(evaluationResults).toEqual(error);
@@ -201,7 +202,7 @@ describe('evaluate', () => {
     const evaluationClient: EvaluationClient = new EvaluationClient();
     pdSchema.input_descriptors[0].schema.push({ uri: "https://www.w3.org/TR/vc-data-model/#types1" });
     vpSimple.verifiableCredential[0].credentialSchema.push({ id: "https://www.w3.org/TR/vc-data-model/#types1" });
-    const evaluationResults = evaluationClient.evaluateWrapper(pdSchema, vpSimple);
+    const evaluationResults = new EvaluationClientWrapper().evaluate(pdSchema, vpSimple, evaluationClient);
     const errorResults = evaluationClient.results.filter(result => result.status === Status.ERROR);
     expect(errorResults.length).toEqual(0);
     expect(evaluationResults).toEqual(success);
@@ -213,7 +214,7 @@ describe('evaluate', () => {
     const vpSimple = getFile('./test/dif_pe_examples/vp/vp_general.json')
     pdSchema.input_descriptors[0].schema[0].uri = "https://business-standards.org/schemas/employment-history.json";
     const evaluationClient: EvaluationClient = new EvaluationClient();
-    const evaluationResults = evaluationClient.evaluateWrapper(pdSchema, vpSimple);
+    const evaluationResults = new EvaluationClientWrapper().evaluate(pdSchema, vpSimple, evaluationClient);
     const errorResults = evaluationClient.results.filter(result => result.status === Status.ERROR);
     const infoResults = evaluationClient.results.filter(result => result.status === Status.INFO);
     expect(errorResults.length).toEqual(7);
@@ -226,7 +227,7 @@ describe('evaluate', () => {
     const vpSimple = getFile('./test/dif_pe_examples/vp/vp_general.json')
     pdSchema.input_descriptors[0].schema[0].uri = "https://eu.com/claims/DriversLicense";
     const evaluationClient: EvaluationClient = new EvaluationClient();
-    const evaluationResults = evaluationClient.evaluateWrapper(pdSchema, vpSimple);
+    const evaluationResults = new EvaluationClientWrapper().evaluate(pdSchema, vpSimple, evaluationClient);
     const errorResults = evaluationClient.results.filter(result => result.status === Status.ERROR);
     const infoResults = evaluationClient.results.filter(result => result.status === Status.INFO);
     expect(errorResults.length).toEqual(5);
@@ -238,7 +239,7 @@ describe('evaluate', () => {
     const pdSchema: PresentationDefinition = getFile('./test/dif_pe_examples/pd/pd-simple-schema-age-predicate.json').presentation_definition;
     const vpSimple = getFile('./test/dif_pe_examples/vp/vp-simple-age-predicate.json');
     const evaluationClient: EvaluationClient = new EvaluationClient();
-    const evaluationResults = evaluationClient.evaluateWrapper(pdSchema, vpSimple);
+    const evaluationResults = new EvaluationClientWrapper().evaluate(pdSchema, vpSimple, evaluationClient);
     expect(evaluationClient.verifiablePresentation.verifiableCredential[0].etc).toEqual(undefined);
     expect(evaluationResults).toEqual(success);
   });
@@ -248,7 +249,7 @@ describe('evaluate', () => {
     const vpSimple = getFile('./test/dif_pe_examples/vp/vp-simple-age-predicate.json');
     const evaluationClient: EvaluationClient = new EvaluationClient();
     delete pdSchema.input_descriptors[0].constraints.limit_disclosure;
-    const evaluationResults = evaluationClient.evaluateWrapper(pdSchema, vpSimple);
+    const evaluationResults = new EvaluationClientWrapper().evaluate(pdSchema, vpSimple, evaluationClient);
     expect(evaluationClient.verifiablePresentation.verifiableCredential[0].etc).toEqual("etc");
     expect(evaluationResults).toEqual(success);
   });
@@ -258,7 +259,7 @@ describe('evaluate', () => {
     const vpSimple = getFile('./test/dif_pe_examples/vp/vp-simple-age-predicate.json');
     const evaluationClient: EvaluationClient = new EvaluationClient();
     pdSchema.input_descriptors[0].constraints.limit_disclosure = Optionality.Preferred;
-    const evaluationResults = evaluationClient.evaluateWrapper(pdSchema, vpSimple);
+    const evaluationResults = new EvaluationClientWrapper().evaluate(pdSchema, vpSimple, evaluationClient);
     expect(evaluationClient.verifiablePresentation.verifiableCredential[0].etc).toEqual("etc");
     expect(evaluationResults).toEqual(success);
   });
@@ -267,7 +268,7 @@ describe('evaluate', () => {
     const pdSchema: PresentationDefinition = getFile('./test/dif_pe_examples/pd/pd-schema-multiple-constraints.json').presentation_definition;
     const vpSimple = getFile('./test/dif_pe_examples/vp/vp-multiple-constraints.json');
     const evaluationClient: EvaluationClient = new EvaluationClient();
-    const evaluationResults = evaluationClient.evaluateWrapper(pdSchema, vpSimple);
+    const evaluationResults = new EvaluationClientWrapper().evaluate(pdSchema, vpSimple, evaluationClient);
     expect(evaluationClient.verifiablePresentation.verifiableCredential[0].birthPlace).toEqual(undefined);
     expect(evaluationResults).toEqual(success);
   });
@@ -277,7 +278,7 @@ describe('evaluate', () => {
     const vpSimple = getFile('./test/dif_pe_examples/vp/vp-simple-age-predicate.json');
     const evaluationClient: EvaluationClient = new EvaluationClient();
     pdSchema.input_descriptors
-    const evaluationResults = evaluationClient.evaluateWrapper(pdSchema, vpSimple);
+    const evaluationResults = new EvaluationClientWrapper().evaluate(pdSchema, vpSimple, evaluationClient);
     expect(evaluationClient.verifiablePresentation.verifiableCredential[0].etc).toEqual(undefined);
     expect(evaluationResults).toEqual(success);
   });
@@ -286,7 +287,7 @@ describe('evaluate', () => {
     const pdSchema: PresentationDefinition = getFile('./test/dif_pe_examples/pd/pd-schema-multiple-constraints.json').presentation_definition;
     const vpSimple = getFile('./test/dif_pe_examples/vp/vp-multiple-constraints.json');
     const evaluationClient: EvaluationClient = new EvaluationClient();
-    const evaluationResults = evaluationClient.evaluateWrapper(pdSchema, vpSimple);
+    const evaluationResults = new EvaluationClientWrapper().evaluate(pdSchema, vpSimple, evaluationClient);
     expect(evaluationClient.verifiablePresentation.verifiableCredential[0].birthPlace).toEqual(undefined);
     expect(evaluationResults).toEqual(success);
   });
